Introduce a Relative Layout that holds both the AztexText and the Toolbar. Expose it to RN side.
